# stock/src/Swagger_biblify_stock_api_original.yaml
...
swagger: "2.0"
info:
  description: |
    This is a sample Biblify stock server .
  version: 1.0.0
  title: Swagger Biblify Stock v1
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /biblify/stock/1.0.0
...
schemes:
- https
- http
paths:
...
  /v1/book/{bookId}:
    get:
      tags:
      - book
      summary: Find book by ID
      description: Returns a single book
      operationId: getBookById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: bookId
        in: path
        description: ID of book to return
        required: true
        type: integer
        format: int64
      security:
      - api_key: []
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Book'
        "400":
          description: Invalid ID supplied
        "404":
          description: Book not found
    put:
      tags:
      - book
      summary: Updates a book in the stock with form data
      operationId: updateBookWithForm
      consumes:
      - application/x-www-form-urlencoded
      - multipart/form-data
      produces:
      - application/json
      - application/xml
      parameters:
      - name: bookId
        in: path
        description: ID of book that needs to be updated
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name of the book
        required: false
        type: string
      - name: amountInStock
        in: formData
        description: Updated amount in stock of the book
        required: false
        type: integer
        format: int64
      - name: amountReserved
        in: formData
        description: Update amount reserved of the book
        required: false
        type: integer
        format: int64
      - name: additionalMetadata
        in: formData
        description: Additional data to pass to server
        required: false
        type: string
      - name: file
        in: formData
        description: file to upload
        required: false
        type: file
      security:
      - biblify_auth:
        - write:books
        - read:books
      responses:
        "200":
          description: Updated successfully
        "404":
          description: Invalid book ID
        "405":
          description: Invalid input
    delete:
      tags:
      - book
      summary: Deletes a book
      operationId: deleteBook
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: bookId
        in: path
        description: Book id to delete
        required: true
        type: integer
        format: int64
      security:
      - biblify_auth:
        - write:books
        - read:books
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Book not found
        "200":
          description: Deleted successfully
...
  /v1/author:
    get:
      tags:
      - author
      summary: Returns existing authors
      description: provide the name to filter results
      operationId: findAuthor
      produces:
      - application/json
      - application/xml
      parameters:
      - name: Name
        in: query
        description: Name of author
        required: false
        type: string
        pattern: ^[a-zA-Z0-9 ]*$
      security:
      - api_key: []
      responses:
        "200":
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Author'
        "400":
          description: Invalid name
        "404":
          description: No author found
    post:
      tags:
      - author
      summary: Add a new author
      operationId: addAuthor
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: author object that needs to be added
        required: true
        schema:
          $ref: '#/definitions/Author'
      security:
      - biblify_auth:
        - write:authors
        - read:authors
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "405":
          description: Invalid input
  /v1/author/{id}:
    get:
      tags:
      - author
      summary: Find author by ID
      description: Returns a single author
      operationId: getAuthorById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of author to return
        required: true
        type: integer
        format: int64
      security:
      - api_key: []
      responses:
        "200":
          description: successful operation
          schema:
            $ref: '#/definitions/Author'
        "400":
          description: Invalid ID supplied
        "404":
          description: author not found
    put:
      tags:
      - author
      summary: Update an existing author
      operationId: updateAuthor
      consumes:
      - application/json
      - application/xml
      - multipart/form-data
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of author to return
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name
        required: false
        type: string
      security:
      - biblify_auth:
        - write:authors
        - read:authors
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Author not found
        "405":
          description: Validation exception
    delete:
      tags:
      - author
      summary: Deletes a author
      operationId: deleteAuthor
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: id
        in: path
        description: author id to delete
        required: true
        type: integer
        format: int64
      security:
      - biblify_auth:
        - write:authors
        - read:authors
      responses:
        "200":
          description: Ok
        "400":
          description: Invalid ID supplied
        "401":
          description: Unauthorized
        "404":
          description: author not found
...
securityDefinitions:
  biblify_auth:
    type: oauth2
    authorizationUrl: http://biblify.swagger.io/oauth/dialog
    flow: implicit
    scopes:
      write:books: modify books
      read:books: read books
      write:authors: modify authors
      read:authors: read authors
      write:orders: write orders
      read:orders: read orders
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
...
  Category:
    type: object
    required:
    - id
    - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Category
  Author:
    type: object
    required:
    - id
    - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    xml:
      name: Author
  Publisher:
    type: object
    required:
    - id
    - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: Scholastic Inc.
    xml:
      name: Publisher
...
  Book:
    type: object
    required:
    - amountInStock
    - amountReserved
    - authors
    - bookId
    - categories
    - name
    properties:
      bookId:
        type: integer
        format: int64
      name:
        type: string
        example: Harry Potter and the Order of the Phoenix
      authors:
        type: array
        items:
          $ref: '#/definitions/Author'
      categories:
        type: array
        items:
          $ref: '#/definitions/Category'
      pictureUrls:
        type: array
        xml:
          name: pictureUrl
          wrapped: true
        items:
          type: string
      edition:
        type: string
        example: The Illustrated Edition (Harry Potter, Book 5)
      publisher:
        $ref: '#/definitions/Publisher'
      date:
        type: string
        format: date
        example: 2017-01-01
      amountInStock:
        type: integer
        format: int64
      amountReserved:
        type: integer
        format: int64
    xml:
      name: Book
...